requirement.txt
crewai
langchain-google-genai
load_dotenv
crewai-tools

langchain
langchain-openai
python-decouple
langchainhub
duckduckgo-search



